+++ src/main/java/com/daemawiki/daemawiki/global/utils/encrypt/KeyConfig.java
@@ -1,4 +1,4 @@
-package com.daemawiki.daemawiki.global.utils.crypto;
+package com.daemawiki.daemawiki.global.utils.encrypt;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
